+++ Cargo.toml
@@ -23,6 +23,7 @@ syntect = { version = "5.0.0", optional = true, default-features = false, featur
 ] }
 
 document-features = { version = "0.2", optional = true }
+comrak = { version = "0.20.0", default-features = false }
 
 [features]
 default = ["load-images"]
