+++ examples/markdown/headers.md
@@ -1,3 +1,5 @@
+
+_aa_ **b** ~~c~~
 # Header 1
 
 ## Header 2
@@ -10,3 +12,17 @@
 
 ###### Header 6
 
+
+
+- a
+    1. a2
+    2. a3
+    3. a4
+- b
+    1. b
+    2. c
+- c
+    - aha
+        - foo
+        - bar
+    - we
